Removed utility.hpp
