+++ src/cpps/parser.hpp
@@ -4,6 +4,7 @@ namespace CPPS::CST {
 
 class Parser
 {
+public:
 };
 
 } // namespace CPPS::CST
