+++ test/unit-tests/cpps/CMakeLists.txt
@@ -17,12 +17,15 @@ set(CPPS_UNIT_TESTS_SOURCES
     cst/node-tests.cpp
     cst/node-variant-tests.cpp
     grammar/boolean-literal-tests.cpp
+    grammar/character-literal-tests.cpp
     grammar/function-modifier-tests.cpp
+    grammar/identifier-tests.cpp
     grammar/keyword-tests.cpp
     grammar/parameter-modifier-tests.cpp
     grammar/pointer-literal-tests.cpp
     grammar/punctuator-tests.cpp
     utility/bump-pointer-allocator-tests.cpp
+    utility/strings-tests.cpp
     check-diagnosis.cpp
     diagnosis-tests.cpp
     lexeme-tests.cpp
@@ -32,7 +35,6 @@ set(CPPS_UNIT_TESTS_SOURCES
     source-tests.cpp
     token-tests.cpp
     tokens-tests.cpp
-    utility-tests.cpp
 )
 
 add_executable(cpps-unit-tests ${CPPS_UNIT_TESTS_INCLUDES} ${CPPS_UNIT_TESTS_SOURCES})
